Ran Astyle to beautify the code. Removed a bunch of legacy code. Fixed a bunch of compiler warnings. Added code to initialize array of polls in sensor.c.
git-svn-id: svn+ssh://svn.ece.msstate.edu/data/www/htdocs/courses/design/2010/wiresoil/svn@61 db668ce9-9f08-4dc0-9949-e28e8fe2f565

+++ wiresoil/common/adc_sample.c
@@ -3,86 +3,89 @@
 #include <stdio.h>
 
 // Samples 1.250V Reference, Calculates Battery Voltage
-float sampleRef(void){
-  uint16 avgADC, i;
-  float f_ref;
-  avgADC = 0;
-  for(i=0;i<32;i++){
-    avgADC = avgADC + convertADC1();
-  }
-  f_ref = (1.25 * 1024.0) / (avgADC / 32.0);
-  printf("Ref: %f\n", f_ref);
-  return f_ref;
+float sampleRef(void)
+{
+    uint16 avgADC, i;
+    float f_ref;
+    avgADC = 0;
+
+    for(i = 0; i < 32; i++)
+    {
+        avgADC = avgADC + convertADC1();
+    }
+
+    f_ref = (1.25 * 1024.0) / (avgADC / 32.0);
+    printf("Ref: %f\n", (double)f_ref);
+    return f_ref;
 }
 
 // Samples Soil Probe
-float sampleSP(float f_bv){
-  uint16 avgADC, i;
-  float f_sp;
-  avgADC = 0;
-  for(i=0;i<32;i++){
-    avgADC = avgADC + convertADC1();
-  }
-  f_sp = ((f_bv / 1024.0 * (avgADC / 32.0)) - 1.250) * 1000.0;
-  printf("Soil Probe: %f\n", f_sp);
-  return f_sp;
+float sampleSP(float f_bv)
+{
+    uint16 avgADC, i;
+    float f_sp;
+    avgADC = 0;
+
+    for(i = 0; i < 32; i++)
+    {
+        avgADC = avgADC + convertADC1();
+    }
+
+    f_sp = ((f_bv / 1024.0 * (avgADC / 32.0)) - 1.250) * 1000.0;
+    printf("Soil Probe: %f\n", (double)f_sp);
+    return f_sp;
 }
 
 // Samples Temperature Indicator
-float sampleTI(float f_bv){
-  uint16 avgADC, i;
-  float f_ti;
-  avgADC = 0;
-  for(i=0;i<32;i++){
-    avgADC = avgADC + convertADC1();
-  }
-  f_ti = ((((f_bv / 1024.0 * (avgADC / 32.0)) * 1000.0)-424.0) / 6.25);
-  printf("Temp: %f\n", f_ti);
-  return f_ti;
+float sampleTI(float f_bv)
+{
+    uint16 avgADC, i;
+    float f_ti;
+    avgADC = 0;
+
+    for(i = 0; i < 32; i++)
+    {
+        avgADC = avgADC + convertADC1();
+    }
+
+    f_ti = ((((f_bv / 1024.0 * (avgADC / 32.0)) * 1000.0) - 424.0) / 6.25);
+    printf("Temp: %f\n", (double)f_ti);
+    return f_ti;
 }
 
-void sampleProbes(FLOAT *af_probeData) {
-  float f_bv;
-  
-  // Calculate Battery Voltage from 1.250V Reference
-  configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN0, 31, ADC_10BIT_FLAG);
-  af_probeData[0].f = f_bv = sampleRef();
-
-  // Sample Soil Probe 1
-  configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN9, 31, ADC_10BIT_FLAG);
-  af_probeData[1].f = sampleSP(f_bv);
-
-  // Sample Soil Probe 2
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN10, 31, ADC_10BIT_FLAG );
-  af_probeData[2].f = sampleSP(f_bv);
-  
-  // Sample Soil Probe 3
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN11, 31, ADC_10BIT_FLAG );
-  af_probeData[3].f = sampleSP(f_bv);
-  
-  // Sample Soil Probe 4
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN12, 31, ADC_10BIT_FLAG );
-  af_probeData[4].f = sampleSP(f_bv);
-
-  // Sample Temp Indicator 1
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN1, 31, ADC_10BIT_FLAG );
-  af_probeData[5].f = sampleTI(f_bv);
-
-  // Sample Temp Indicator 2
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN2, 31, ADC_10BIT_FLAG );
-  af_probeData[6].f = sampleTI(f_bv);
-
-  // Sample Temp Indicator 3
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN3, 31, ADC_10BIT_FLAG );
-  af_probeData[7].f = sampleTI(f_bv);
-  
-  // Sample Temp Indicator 4
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN4, 31, ADC_10BIT_FLAG );
-  af_probeData[8].f = sampleTI(f_bv);
-
-  // Sample Temp Indicator 5
-  configADC1_ManualCH0( ADC_CH0_POS_SAMPLEA_AN5, 31, ADC_10BIT_FLAG );
-  af_probeData[9].f = sampleTI(f_bv);
+void sampleProbes(FLOAT *af_probeData)
+{
+    float f_bv;
+    // Calculate Battery Voltage from 1.250V Reference
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN0, 31, ADC_10BIT_FLAG);
+    af_probeData[0].f = f_bv = sampleRef();
+    // Sample Soil Probe 1
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN9, 31, ADC_10BIT_FLAG);
+    af_probeData[1].f = sampleSP(f_bv);
+    // Sample Soil Probe 2
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN10, 31, ADC_10BIT_FLAG);
+    af_probeData[2].f = sampleSP(f_bv);
+    // Sample Soil Probe 3
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN11, 31, ADC_10BIT_FLAG);
+    af_probeData[3].f = sampleSP(f_bv);
+    // Sample Soil Probe 4
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN12, 31, ADC_10BIT_FLAG);
+    af_probeData[4].f = sampleSP(f_bv);
+    // Sample Temp Indicator 1
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN1, 31, ADC_10BIT_FLAG);
+    af_probeData[5].f = sampleTI(f_bv);
+    // Sample Temp Indicator 2
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN2, 31, ADC_10BIT_FLAG);
+    af_probeData[6].f = sampleTI(f_bv);
+    // Sample Temp Indicator 3
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN3, 31, ADC_10BIT_FLAG);
+    af_probeData[7].f = sampleTI(f_bv);
+    // Sample Temp Indicator 4
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN4, 31, ADC_10BIT_FLAG);
+    af_probeData[8].f = sampleTI(f_bv);
+    // Sample Temp Indicator 5
+    configADC1_ManualCH0(ADC_CH0_POS_SAMPLEA_AN5, 31, ADC_10BIT_FLAG);
+    af_probeData[9].f = sampleTI(f_bv);
 }
 
 /*void NormalMode(void) {
